Updated makefile and initial sources placement
* Added BOOST_PREFIX compile options to qmake and make commands
* Added fix from GR-ABM-ODE-3D for initial sources placement

git-svn-id: svn://innoculant.micro.med.umich.edu/dev/gr2d/GR-ABM-ODE@1051 88ce85bb-8564-4ae6-884f-ab7503230054

+++ simulation/grstat.cpp
@@ -148,7 +148,7 @@ void GrStat::updateAgentStatistics(Agent* a)
       _macIntMtbStats[pMac->getState()](pMac->getIntMtb());
       if(pMac->getState() == MAC_INFECTED || pMac->getState() == MAC_CINFECTED){
         assert(pMac->getIntMtb() < _PARAM(PARAM_MAC_THRESHOLD_BURST_CI_INTMTB));
-        assert((int) pMac->getIntMtb() < _intMtbFreqSize);
+        assert(pMac->getIntMtb() < _intMtbFreqSize);
 
         _intMtbFreq[int(pMac->getIntMtb())]++;
       }
